refactor: re-use add_turbo_fish function
